LazyEventBusInvocationHandler for eventBus proxy
