+++ platform/dolphin-platform-server-javaee/src/main/java/com/canoo/dp/impl/server/javaee/CdiBeanFactory.java
@@ -16,22 +16,21 @@
 package com.canoo.dp.impl.server.javaee;
 
 import com.canoo.dolphin.BeanManager;
-import com.canoo.platform.server.RemotingContext;
-import com.canoo.platform.server.binding.PropertyBinder;
-import com.canoo.impl.server.context.DolphinContext;
-import com.canoo.impl.server.context.DolphinContextProvider;
-import com.canoo.impl.server.context.RemotingContextImpl;
-import com.canoo.platform.server.event.DolphinEventBus;
 import com.canoo.impl.platform.core.Assert;
 import com.canoo.impl.server.bootstrap.PlatformBootstrap;
 import com.canoo.impl.server.client.ClientSessionProvider;
+import com.canoo.impl.server.context.DolphinContext;
+import com.canoo.impl.server.context.DolphinContextProvider;
+import com.canoo.impl.server.context.RemotingContextImpl;
+import com.canoo.impl.server.event.LazyEventBusInvocationHandler;
+import com.canoo.platform.server.RemotingContext;
+import com.canoo.platform.server.binding.PropertyBinder;
 import com.canoo.platform.server.client.ClientSession;
+import com.canoo.platform.server.event.DolphinEventBus;
 import com.canoo.platform.server.javaee.ClientScoped;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
 
 /**
@@ -74,20 +73,7 @@ public class CdiBeanFactory {
     @Produces
     @ApplicationScoped
     public DolphinEventBus createEventBus() {
-        return (DolphinEventBus) Proxy.newProxyInstance(getClass().getClassLoader(), new Class[]{DolphinEventBus.class}, new InvocationHandler() {
-            @Override
-            public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-                DolphinEventBus instance = PlatformBootstrap.getServerCoreComponents().getInstance(DolphinEventBus.class);
-                if (instance != null) {
-                    return method.invoke(instance, args);
-                }
-                if (method.getName().equals("publish")) {
-                    return null;
-                } else {
-                    throw new IllegalStateException("Subscription can only be done from Dolphin Context!");
-                }
-            }
-        });
+        return (DolphinEventBus) Proxy.newProxyInstance(getClass().getClassLoader(), new Class[]{DolphinEventBus.class}, new LazyEventBusInvocationHandler());
     }
 
     @Produces
